+++ src/components/common/Header.js
@@ -25,12 +25,8 @@ const ServiceMemberMenuItems = [
   },
   {
     label: 'Career Counseling',
-    path: '/serviceMember/degreePathwaysList',
+    path: '/serviceMember/careerCounseling',
   },
-  // {
-  //   label: 'Career Plans',
-  //   path: '/serviceMember/careerPlans',
-  // },
   {
     label: 'Quick Links',
     path: '/quickLinks',
